create new avro builder file
